fix: React Routerコンテキストの階層を修正してE2Eテストエラーを解消
- AuthProviderをRouterProvider内に配置してuseNavigateフックのエラーを修正
- AppRoutes.tsxを削除し、ルーティング定義をAppProviders.tsxに統合
- Outlet利用でAuthProviderが全ルートに適用される構造に変更
- SignInPageコンポーネントが正しくレンダリングされるように修正
